requirements.txt
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.9.1
python-dotenv==0.21.0
requests==2.28.1
toml==0.10.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
